# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "Register request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.LoginResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.UserResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user notifications",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Only unread",
                        "name": "unread",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.NotificationListResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me/notifications/read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Mark notifications as read",
                "parameters": [
                    {
                        "description": "Notification IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.MarkReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/users/me/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.UserStats"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me/watchlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user watchlist",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.WatchlistWithSkin"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/me/watchlist/{skin_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Add skin to watchlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skin ID",
                        "name": "skin_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Watchlist"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Remove skin from watchlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skin ID",
                        "name": "skin_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "entity.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entity.UserResponse"
                }
            }
        },
        "entity.MarkReadRequest": {
            "type": "object",
            "required": [
                "notification_ids"
            ],
            "properties": {
                "notification_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "entity.Notification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "description": "JSON данные",
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/entity.NotificationPriority"
                },
                "read_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/entity.NotificationType"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.NotificationListResponse": {
            "type": "object",
            "properties": {
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Notification"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "unread": {
                    "type": "integer"
                }
            }
        },
        "entity.NotificationPriority": {
            "type": "string",
            "enum": [
                "low",
                "normal",
                "high",
                "urgent"
            ],
            "x-enum-varnames": [
                "PriorityLow",
                "PriorityNormal",
                "PriorityHigh",
                "PriorityUrgent"
            ]
        },
        "entity.NotificationType": {
            "type": "string",
            "enum": [
                "price_drop",
                "price_increase",
                "target_reached",
                "skin_discovered",
                "system_alert",
                "watchlist_update"
            ],
            "x-enum-varnames": [
                "TypePriceDrop",
                "TypePriceIncrease",
                "TypeTargetReached",
                "TypeSkinDiscovered",
                "TypeSystemAlert",
                "TypeWatchlistUpdate"
            ]
        },
...
        "entity.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "entity.Skin": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "current_price": {
                    "type": "number"
                },
                "highest_price": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "last_updated": {
                    "type": "string"
                },
                "lowest_price": {
                    "type": "number"
                },
                "market_hash_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_change_24h": {
                    "type": "number"
                },
                "price_change_7d": {
                    "type": "number"
                },
                "quality": {
                    "type": "string"
                },
                "rarity": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "volume_24h": {
                    "type": "integer"
                },
                "weapon": {
                    "type": "string"
                }
            }
        },
...
        "entity.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/entity.UserRole"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "entity.UserRole": {
            "type": "string",
            "enum": [
                "user",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleAdmin"
            ]
        },
        "entity.UserStats": {
            "type": "object",
            "properties": {
                "active_alerts": {
                    "type": "integer"
                },
                "member_since": {
                    "type": "string"
                },
                "total_notifications": {
                    "type": "integer"
                },
                "total_watchlist": {
                    "type": "integer"
                },
                "unread_notifications": {
                    "type": "integer"
                }
            }
        },
        "entity.Watchlist": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "notify_on_drop": {
                    "type": "boolean"
                },
                "notify_on_price": {
                    "type": "boolean"
                },
                "skin_id": {
                    "type": "string"
                },
                "target_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.WatchlistWithSkin": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "notify_on_drop": {
                    "type": "boolean"
                },
                "notify_on_price": {
                    "type": "boolean"
                },
                "skin": {
                    "$ref": "#/definitions/entity.Skin"
                },
                "skin_id": {
                    "type": "string"
                },
                "target_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}